requirements.txt
delphin==0.0.2
discord.py==2.0.1
emoji==1.6.3
HarfPy==0.9
matplotlib==3.5.1
numpy==1.22.2
pickleDB==0.9.2
PyDelphin==1.6.0
Qahirah==1.1
